feat: refactor wallet connect button
