+++ config/wagmi/config.ts
@@ -1,10 +1,10 @@
-import { createConfig, http } from "wagmi";
-import { injected } from "@wagmi/connectors";
+import { http, createConfig } from "wagmi";
 import { base } from "wagmi/chains";
+import { injected, metaMask, safe } from "wagmi/connectors";
 
-export const wagmiConfig = createConfig({
+export const config = createConfig({
   chains: [base],
-  connectors: [injected()],
+  connectors: [injected(), metaMask(), safe()],
   transports: {
     [base.id]: http(),
   },
