+++ package.json
@@ -22,12 +22,12 @@
     "@types/react-dom": "^18.2.17",
     "@vitejs/plugin-react": "^4.2.1",
     "dotenv": "^16.3.1",
-    "typescript": "^5.2.2",
     "viem": "^2.0.0",
     "vite": "^5.0.8",
     "wagmi": "^2.0.3"
   },
   "devDependencies": {
+    "typescript": "^5.2.2",
     "@typescript-eslint/eslint-plugin": "^6.14.0",
     "@typescript-eslint/parser": "^6.14.0",
     "eslint": "^8.55.0",
