+++ src/App.tsx
@@ -1,5 +1,5 @@
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
-import { wagmiConfig } from "../config/wagmi/config";
+import { config } from "../config/wagmi/config";
 import { WagmiProvider } from "wagmi";
 import ResponsiveAppBar from "./components/ResponsiveAppBar";
 import InfoPanel from "./components/InfoPanel";
@@ -8,7 +8,7 @@ const queryClient = new QueryClient();
 
 function App() {
   return (
-    <WagmiProvider config={wagmiConfig}>
+    <WagmiProvider config={config}>
       <QueryClientProvider client={queryClient}>
         <ResponsiveAppBar />
         <InfoPanel />
